adapter: Removes report handler and call AuthRep in check template
Due to some undesired caching behaviour[1] at the Istio/Envoy layer, this change
moves to calling 3scale AuthRep endpoiont in the Authorization template handler.

Caching has been disabled in the 1.0.x release[2] (fixing above issue) of Istio at the time of this change.
We therefore cannot depend on it's behaviour for release and so temporarily defaulting validation values
to zero to disable all caching.

[1]https://github.com/3scale/istio-integration/issues/29
[2]https://github.com/istio/istio/issues/9596